voice-agent/orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
websockets>=13.1
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.9.2
prometheus-client>=0.20.0
edge-tts>=6.1.0
audioop-lts>=0.2.1;python_version>="3.13"